Bug/FP-657 use system name instead of my data (#199)
* Use system name in search bar and in copy/move/select modals

* Improve import statement

* Update upload modal to use proper system name

* Fix tests

* Add doc strings

* Add test for finding display name of system
