+++ client/src/components/DataFiles/DataFilesBreadcrumbs/DataFilesBreadcrumbs.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import { v4 as uuidv4 } from 'uuid';
 import './DataFilesBreadcrumbs.scss';
-import { getSystemName } from 'utils/jobsUtil';
+import { findSystemDisplayName } from 'utils/systems';
 
 const BreadcrumbLink = ({ api, scheme, system, path, children, section }) => {
   const dispatch = useDispatch();
@@ -70,13 +70,6 @@ const DataFilesBreadcrumbs = ({
   const paths = [];
   const pathComps = [];
   const systemList = useSelector(state => state.systems.systemList);
-  const privateSystemName = () => {
-    const matchingSystem = systemList.find(s => s.system === system);
-    if (matchingSystem) {
-      return matchingSystem.name;
-    }
-    return getSystemName(system);
-  };
 
   path
     .split('/')
@@ -91,7 +84,7 @@ const DataFilesBreadcrumbs = ({
   const root = (() => {
     switch (scheme) {
       case 'private':
-        return privateSystemName();
+        return findSystemDisplayName(systemList, system);
       case 'community':
         return 'Community Data';
       default:
