+++ client/src/utils/jobsUtil.test.js
@@ -1,6 +1,5 @@
 import {
   getAllocatonFromDirective,
-  getSystemName,
   getJobDisplayInformation,
   getOutputPathFromHref,
   isTerminalState,
@@ -30,10 +29,6 @@ describe('jobsUtil', () => {
     expect(
       getJobDisplayInformation(jobDetailFixture, appDetailFixture)
     ).toEqual(jobDetailDisplayFixture);
-  });
-
-  it('get system name from host', () => {
-    expect(getSystemName('stampede2.tacc.utexas.edu')).toEqual('Stampede2');
   });
 
   it('get output path from _links.archiveData.href', () => {
